Add IAttackSequence interface, update player controls and character controller system
The IAttackSequence interface for handling different types of attack sequences was added. Player Controls.inputactions was updated with new button interactions and actions. The surface contact system was removed from the character controller system, and modifications to methods in RigidbodyFirstPersonController.cs were made for optimizing the character's movements.

Adding ChargedAttacks and Warp Strikes